UserController Test of GetUser() was successful
